Melhorias nas implementações do modulo wtom-view-web (#1)
